Fix processing of chain operator while accessing class member function.
In order to inform the ETS codegen about optional functin call it is required
to set the 'optional_' field to true while processing the call expression at the
'parser/ETSparser.cpp' file.

Also, it was found that for optional call to class member function the detected
chain expression flag is not count (while for member field it is processed
correctly) by the ETS codegen.

This commitis to provide the info with respect to detected chain expression
operand to subsequent parsing of function call while calling ParseCallExpression()
and update the bytecode generation for class member function call with respect to
actual value of the 'optional_' field.

Signed-off-by: Vadim Lomovtsev <vadim.lomovtsev@auriga.com>