Fix casting const union
Issue: https://gitee.com/openharmony/arkcompiler_ets_frontend/issues/I9MV5J

Reason
When casting an expression of union type to one of the types of the union, the compilation should be successful and can not cause 'std::bad_alloc'.

Description
When setting the TsType to the variable, it is not necessary to assign the constness.

Tests
ninja es2panda_tests
bash ets_testrunner.sh --cts -r -rt

Signed-off-by: daizihan <daizihan@huawei.com>