Fixing module according to the standard
Issue: https://gitee.com/openharmony/arkcompiler_ets_frontend/issues/I9V12U

The naming conventions for modules are not clear, in some places the separate module
is treated as a package module and there are unnecessary checks also. This issue aims to fix that.

In addition to using these two modules (package and separate), there is a third special case:
if the '--ets-module' compiler option does not added and the program is a separate module and also an entry program.
Currently, the compiler does not treat them as a separate module in the same way (should do),
and also omit the module name in the record table (will keep this for now, should be reconsidered)

Signed-off-by: Robert Sipka <robert.sipka@huawei.com>