Modify the returntype assignability logic
Issue:
https://gitee.com/openharmony/arkcompiler_ets_frontend/issues/IAZXOP

Description:
When the returntype of override is primitivetype, false is returned when source returntype and target returntype are different. Because primitivetype has no returntypeannotation, obtaining the returntypeannotation directly results in a segment fault.

When the returntype of override is object, and source returntype and target returntype cannot match, the system checks whether target is typeparameter. If no, false is returned. Getting the returntypeannotation directly causes the segment fault.

Tested-by:
ninja tests (passed)
ets_testrunner (passed)

Signed-off-by: zengzengran <zengran2@huawei.com>##